backend/requirements.txt
Django>=6.0.1
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=24.1
firebase-admin>=6.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
Pillow>=10.0.0

